apps/server/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.9

google-adk[database]==1.14.0
yfinance==0.2.65
psutil==7.0.0
litellm==1.77.0
google-generativeai==0.8.5
python-dotenv==1.1.1